// packages/opencrvs-to-dci-builder/src/registry-core-api.d.ts
export interface components {
// ...
  responses: {
    /** @description HTTP layer error details */
    HttpErrorResponse: {
      content: {
        "application/json": {
          error?: {
            /** @description error code */
            code?: string;
            /** @description error message */
            message?: string;
          };
        };
      };
    };
    /** @description Acknowledgement of message received after successful validation of message and signature */
    Response: {
      content: {
        "application/json": {
          message?: {
            ack_status?: components["schemas"]["Ack"];
            timestamp?: components["schemas"]["DateTime"];
            error?: components["schemas"]["Error"];
          };
        };
      };
    };
  };
// ...
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {
// ...
  post_reg_sync_search: {
    requestBody: {
      content: {
        "application/json": {
          signature?: components["schemas"]["MsgSignature"];
          header: components["schemas"]["MsgHeader_V1.0.0"] & {
            /** @enum {unknown} */
            action?: "search";
          };
          /** @description The search data using which registry search to be performed */
          message: components["schemas"]["SearchRequest"];
          //  | components["schemas"]["EncryptedMessage"];
        };
      };
    };
    responses: {
      401: components["responses"]["HttpErrorResponse"];
      403: components["responses"]["HttpErrorResponse"];
      500: components["responses"]["HttpErrorResponse"];
      /** @description Registry search response */
      default: {
        content: {
          "application/json": {
            signature?: components["schemas"]["MsgSignature"];
            header: components["schemas"]["MsgCallbackHeader_V1.0.0"] & {
              /** @enum {unknown} */
              action?: "on-search";
            };
            message?: components["schemas"]["SearchResponse"];
            //  | components["schemas"]["EncryptedMessage"];
          };
        };
      };
    };
  };
// ...
}